tests/wpt/mozilla/tests/input-events/sequential-focus-navigation-across-documents-002.html

<!DOCTYPE html>
<title>Sequential focus navigation works across sandboxed iframe boundaries</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />

<input class="text" id="input1">
<iframe id="frame" sandbox srcdoc="<input id=frameinput1><input id=frameinput2>"></iframe>
<input class="text" id="input2">

<script>
promise_test(async test => {
    let tabKey = "\uE004";
    let shiftKey = "\uE008";

    input1.focus();
    assert_equals(document.activeElement, input1);

    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();
    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();
    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();

    assert_equals(document.activeElement, input2);

    // Now traverse the sequential focus order backwards.
    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    assert_equals(document.activeElement, input1);

});
</script>
